More refactoring, renaming, architecture cleanup
