+++ src/components/AnnotationCommands/AnnotationCommands.tsx
@@ -1,7 +1,9 @@
 import { useState } from 'react';
 import { Braces, Spline, Tags } from 'lucide-react';
+import { EntityType, Tag } from '@/model';
 import { useDataModel } from '@/store';
 import { Button } from '@/ui/Button';
+import { EntityTypeEditor } from '../EntityTypeEditor';
 import {
   Command,
   CommandEmpty,
@@ -11,8 +13,6 @@ import {
   CommandList,
   CommandSeparator
 } from '@/ui/Command';
-import { EntityTypeDialog } from '../EntityTypeDialog';
-import { EntityType, Tag } from '@/model';
 
 interface AnnotationCommandProps {
 
@@ -93,11 +93,11 @@ export const AnnotationCommands = (props: AnnotationCommandProps) => {
         </CommandList>
 
         <div className="p-1 pt-1.5 border-t flex justify-end text-muted-foreground">
-          <EntityTypeDialog>
+          <EntityTypeEditor>
             <Button variant="ghost" className="text-xs h-8 px-2 rounded-sm mr-2">
               <Braces className="h-3.5 w-3.5 mr-1" /> Create new entity
             </Button>
-          </EntityTypeDialog>
+          </EntityTypeEditor>
 
           <Button variant="ghost" className="text-xs h-8 px-2 rounded-sm">
             <Spline className="h-3.5 w-3.5 mr-1" /> Create new relation
