--- src/components/EntityTypeDialog/index.ts
@@ -1,1 +1,0 @@
-export * from './EntityTypeDialog';
