+++ src/components/EntityTypeEditor/Editor/PropertyDefinitions/Properties.tsx
