+++ src/components/EntityTypeEditor/Editor/PropertyDefinitions/PropertiesActions.tsx
