+++ src/components/EntityTypeEditor/Editor/PropertyDefinitions/PropertyEditor.tsx
