+++ src/components/EntityTypeEditor/Editor/PropertyDefinitions/PropertyEditorDialog.tsx
