+++ src/components/EntityTypeEditor/Editor/PropertyDefinitions/moveArrayItem.ts
